chore: update copy based on PR review
